Change all frame offsets to i64 microseconds.

+++ exopticon/src/cworkers/mpack_frame.h
@@ -5,7 +5,7 @@
 struct FrameMessage {
         uint8_t *jpeg;
         int32_t jpeg_size;
-        int32_t offset;
+        int64_t offset;
         int32_t unscaled_height;
         int32_t unscaled_width;
 };
